requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0
langchain-google-genai>=1.0.0
langchain-huggingface
streamlit>=1.28.0
pydantic>=2.0.0
composio-core>=0.5.0
python-dotenv>=1.0.0
httpx>=0.24.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.3
google-api-python-client>=2.80.0
google-auth-httplib2>=0.1.0
